pyproject.toml
[project]
name = "cooursework-6-drf"
version = "0.1.0"
description = ""
authors = [
    {name = "Kovkov-D-S",email = "158774590+Kovkov-D-S@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "setuptools (>=80.3.1,<81.0.0)",
    "django (>=5.2,<6.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "ipython (>=9.2.0,<10.0.0)",
    "redis (>=6.0.0,<7.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "flake8 (>=7.2.0,<8.0.0)",
    "django-countries (>=7.6.1,<8.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "generics (>=7.0.0,<8.0.0)",
    "celery (>=5.5.2,<6.0.0)",
    "django-timezone-field (>=7.1,<8.0)",
    "django-celery-beat (>=2.8.0,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "coverage (>=7.8.0,<8.0.0)",
    "eventlet (>=0.39.1,<0.40.0)",
    "gunicorn (>=23.0.0,<24.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
